Updated with New Supervisor Setup
The Supervisor.Spec module that contains the worker function was deprecated in
Elixir 1.6: https://hexdocs.pm/elixir/v1.6/Supervisor.Spec.html

Defining child_spec/1 in database will allow us to spawn the DB with the
new child definitions. It's included in the macro so all modules that
use database will have this function as well.

The README is updated to correlate with the changes.