Fixed argument bug in run_with_psql Command
Attempting to use run_with_psql with the options for username and
password would fail before due to a test using Enum.member?/2, when in
fact the opts datatype was a Keyword.
This has now been updated to use Keyword.has_key?/2.
In addition, passing in the password as an argument was not a valid
command.
This behavior has been updated to use environment variables to run the
psql command with the password. The username variable is also passed in
this way.
These use the PGPASSWORD and PGUSER environment variables as found here:
https://www.postgresql.org/docs/current/libpq-envars.html